book-store/src/main/java/com/infinitum/model/dtos/AuthorDTO.java
package com.infinitum.model.dtos;

public class AuthorDTO {
    private long id;
    private String firstName;
    private String lastName;

    public AuthorDTO() {
    }

    public AuthorDTO(long id, String firstName, String lastName) {
        this.id = id;
        this.firstName = firstName;
        this.lastName = lastName;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }
}
